impl storage abstraction
