+++ web-api/tests/cmd/assets/config.toml
@@ -5,13 +5,13 @@ file = "foo/bar.log"
 bind_to = "127.0.1.1:3040"
 max_body_size = 64
 
-[elastic]
+[storage.elastic]
 url = "http://localhost:3219"
 user = "tic"
 password = "can not be tested"
 index_name = "other_index"
 
-[db]
+[storage.postgres]
 base_url = "postgres://user:pw@localhost:5432/xayn"
 port = 42
 user = "postgres"
