+++ web-api/tests/config.rs
@@ -58,7 +58,7 @@ fn test_loading_configs() {
 fn test_loading_config_with_env_overrides() {
     with_env_guard(
         [
-            ("XAYN_WEB_API__DB__PORT", "3532"),
+            ("XAYN_WEB_API__STORAGE__POSTGRES__PORT", "3532"),
             ("XAYN_WEB_API__NET__MAX_BODY_SIZE", "4422"),
         ],
         || {
